requirements.txt
Pillow==2.7.0